+++ lib/tcr/actions/tcr_action_run.sh
@@ -12,6 +12,9 @@ TCR_ACTION_RUN_PHASE_TEST="$TCR_TEST_PHASE"
 TCR_ACTION_RUN_PHASE_COMMIT="$TCR_COMMIT_PHASE"
 TCR_ACTION_RUN_PHASE_REVERT="$TCR_REVERT_PHASE"
 
+TCR_NOTIFY_SUCCESS_PHASE="notify_success|Notify|Notifying about success...|Successfully notified.|Notifying failed."
+TCR_NOTIFY_FAILURE_PHASE="notify_failure|Notify|Notifying about failure...|Successfully notified.|Notifying failed."
+
 tcr_action_run() {
     if ! tcr_is_enabled; then
         tcr_error_raise "$TCR_ERROR_TCR_NOT_ENABLED"
@@ -24,38 +27,24 @@ tcr_action_run() {
 
     tcr_print_event "Starting TCR run..."
 
-    execution_phase="$TCR_ACTION_RUN_PHASE_BUILD"
-    execute_phase "$execution_phase"
+    execute_phase "$TCR_ACTION_RUN_PHASE_BUILD"
     phase_error_code=$?
 
-    if is_success "$phase_error_code"; then
-        prev_phase="$execution_phase"
-        prev_error_code="$phase_error_code"
-        
-        execution_phase="$TCR_ACTION_RUN_PHASE_TEST"
-        execute_phase "$execution_phase"
+    if is_success "$phase_error_code"; then        
+        execute_phase "$TCR_ACTION_RUN_PHASE_TEST"
         phase_error_code=$?
 
         if is_success "$phase_error_code"; then
-            prev_phase="$execution_phase"
-            prev_error_code="$phase_error_code"
-
-            execution_phase="$TCR_ACTION_RUN_PHASE_COMMIT"
-            execute_phase "$execution_phase"
-            phase_error_code=$?
+            execute_phase "$TCR_ACTION_RUN_PHASE_COMMIT"
         else            
-            prev_phase="$execution_phase"
-            prev_error_code="$phase_error_code"
-
-            execution_phase="$TCR_ACTION_RUN_PHASE_REVERT"
-            execute_phase "$execution_phase"
-            phase_error_code=$?
-
-            if is_success "$phase_error_code"; then
-                execution_phase="$prev_phase"
-                phase_error_code="$prev_error_code"
-            fi
+            execute_phase "$TCR_ACTION_RUN_PHASE_REVERT"
         fi
+    fi
+
+    if is_success "$phase_error_code"; then
+        execute_phase "$TCR_NOTIFY_SUCCESS_PHASE"
+    else
+        execute_phase "$TCR_NOTIFY_FAILURE_PHASE"
     fi
 
     tcr_print_event "TCR run completed."
@@ -80,6 +69,12 @@ command_for_phase() {
             ;;
         "$TCR_ACTION_RUN_PHASE_COMMIT")
             echo "$TCR_COMMIT_CMD"
+            ;;
+        "$TCR_NOTIFY_SUCCESS_PHASE")
+            echo "$TCR_NOTIFICATION_SUCCESS_CMD"
+            ;;
+        "$TCR_NOTIFY_FAILURE_PHASE")
+            echo "$TCR_NOTIFICATION_FAILURE_CMD"
             ;;
     esac
 }